phase-2/pyproject.toml
[project]
name = "phase-2-todo-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "email-validator>=2.3.0",
    "fastapi>=0.128.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
